[wip]: set up root cmd
